chore: abstract dot TransferBuilder
Create an abstract class out of TransferBuilder so that parachains
can extend this transferBuilder and override buildTransaction() if for
some reason the way to create a Native Transfer transaction for a
parachain differs but moonbeam and acala are the exact same.

TICKET: BG-53437
