+++ modules/sdk-coin-dot/src/lib/index.ts
@@ -14,4 +14,5 @@ export { WithdrawUnstakedBuilder } from './withdrawUnstakedBuilder';
 export { ClaimBuilder } from './claimBuilder';
 export { TransactionBuilderFactory } from './transactionBuilderFactory';
 export { SingletonRegistry } from './singletonRegistry';
+export { NativeTransferBuilder } from './nativeTransferBuilder';
 export { Interface, Utils };
